third_party/WebKit/LayoutTests/fast/html/imports/import-master-expected.txt

PASS master.import.querySelector('h1').innerHTML is 'Importing Master'
PASS master.import.querySelector('#master').import is master.import
PASS successfullyParsed is true

TEST COMPLETE
Importing Master
